Add the ability to make individual charts
Add one demo chart
